+++ pkg/plugin/testdata/authorized-type.yml
