+++ pkg/plugin/testdata/broken-config.yml
