+++ pkg/plugin/testdata/default-type.yml
