+++ pkg/plugin/testdata/unauthorized-type.yml
